Re-organize package
