+++ setup.cfg
@@ -27,8 +27,7 @@ classifiers =
 	Topic :: Scientific/Engineering
 
 [options]
-packages = find:
-py_modules = xpartition
+packages = xpartition
 zip_safe = False  # https://mypy.readthedocs.io/en/latest/installed_packages.html
 include_package_data = True
 python_requires = >=3.6
@@ -52,6 +51,7 @@ ignore =
 exclude = 
 	.eggs
 	doc
+	__init__.py
 
 [bumpversion:file:xpartition.py]
 search = __version__ = "{current_version}"
