+++ xpartition/__init__.py
@@ -1,0 +1,2 @@
+from . import xpartition
+from .xpartition import __version__, zeros_like, PartitionMapper
